rename android hardware
