+++ common/hardware_android.py
